Fix up test tests.
