+++ service-front/app/src/Common/templates/partials/govuk_error.html.twig
@@ -10,7 +10,7 @@
                     {% for id, messages in errors %}
                         {% for message in messages %}
                             <li>
-                                <a href="#{%- if id >= 0 -%}{{ id }}{%- endif -%}">{{ message | raw }}</a>
+                                <a href="#{%- if id >= 0 -%}{{ id }}{%- endif -%}">{{ message | trans | raw }}</a>
                             </li>
                         {% endfor %}
                     {% endfor %}
@@ -25,7 +25,7 @@
         {# ... display the errors for a single field #}
         {%- for error in errors -%}
             <span id="{{ id }}-error" class="govuk-error-message">
-                <span class="govuk-visually-hidden">{% trans %}Error:{% endtrans %}</span> {{ error | raw }}
+                <span class="govuk-visually-hidden">{% trans %}Error:{% endtrans %}</span> {{ error | trans | raw }}
             </span>
         {%- endfor -%}
     {% endif %}
