+++ service-front/app/test/CommonTest/View/Twig/TranslationExtensionFactoryTest.php
@@ -4,13 +4,14 @@ declare(strict_types=1);
 
 namespace CommonTest\View\Twig;
 
-use Common\View\Twig\SymfonyTranslationExtensionFactory;
+use Acpr\I18n\NodeVisitor\TranslationNodeVisitor;
+use Acpr\I18n\TranslationExtension;
+use Acpr\I18n\TranslatorInterface;
+use Common\View\Twig\TranslationExtensionFactory;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
-use Symfony\Bridge\Twig\Extension\TranslationExtension;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
-class SymfonyTranslationExtensionFactoryTest extends TestCase
+class TranslationExtensionFactoryTest extends TestCase
 {
     /** @test */
     public function it_creates_an_extension(): void
@@ -18,8 +19,10 @@ class SymfonyTranslationExtensionFactoryTest extends TestCase
         $containerProphecy = $this->prophesize(ContainerInterface::class);
         $containerProphecy->get(TranslatorInterface::class)
             ->willReturn($this->prophesize(TranslatorInterface::class)->reveal());
+        $containerProphecy->get(TranslationNodeVisitor::class)
+            ->willReturn(new TranslationNodeVisitor()); // Visitor is marked final
 
-        $factory = new SymfonyTranslationExtensionFactory();
+        $factory = new TranslationExtensionFactory();
 
         $instance = $factory($containerProphecy->reveal());
